+++ packages/coding-agent/src/core/tools/index.ts
@@ -22,7 +22,6 @@ export {
 	editToolDefinition,
 } from "./edit.js";
 export { withFileMutationQueue } from "./file-mutation-queue.js";
-export { expandPath, resolveToCwd } from "./path-utils.js";
 export {
 	createFindTool,
 	createFindToolDefinition,
@@ -53,6 +52,7 @@ export {
 	lsTool,
 	lsToolDefinition,
 } from "./ls.js";
+export { expandPath, resolveToCwd } from "./path-utils.js";
 export {
 	createReadTool,
 	createReadToolDefinition,